Revert extract
